Translation

+++ type-checker/src/ast_nodes.rs
@@ -5,7 +5,7 @@ use strum_macros::EnumIter;
 pub type Id = String;
 use from_variants::FromVariants;
 
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, EnumIter, Copy)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, EnumIter, Copy, Hash, Eq)]
 pub enum AtomicTypeEnum {
     INT,
     BOOL,
@@ -223,7 +223,7 @@ pub fn Var(id: &str) -> GenericVariable {
 #[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
 pub struct ElementAccess {
     pub expression: Box<Expression>,
-    pub index: u32,
+    pub index: usize,
 }
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
